deployment/requirements.txt
aws-cdk-lib==2.148.0
constructs>=10.0.0,<11.0.0
aws-wsgi
flask
aws_lambda_python_alpha
boto3
mlflow
pandas
sklearn
